Better styling on the landing page
We simplify the rendered markup and some of the logic of the render function.
This allows us to also simplify the CSS, by mostly using the capabilities of
flexbox to have the same UI as we had before (no absolute positionning).
We also add some custom properties so the styling is more consistent.