Fix tests that were failing due to the change to functional components.
The main failure was the one testing that the Rep component was picking
up the expected Rep (Grip, TextNode, ...) component given a specific
object.
The test is modified to use the `getRep` function, that the Rep (from
rep.js) calls to render the expected component, since the shallow render
would only gives us the type returned by the function, and not the
function name.