Removing the stack-trace module
The stack-trace module was only in devtools-core for the benefit of the
console.html project which isn't actively pursued right now. All console
work happens in the mozilla-central repository nowadays.
This module is also in mozilla-central, so there's no sense keeping it
here too. Removing it gets rid of some code duplication.