Use preview.isConnected on node to check if it is in DOM tree.
With https://bugzilla.mozilla.org/show_bug.cgi\?id\=1347490 we added a property to the
Node ObjectActor which states if the node is in the DOM tree or not.
This we can simplify our code and get rid of the attachedActorIds thing we were using
previously.
Now consumer would only have to pass an onInspectIconClick to handle the icon.