Don't wrap PropRep elements in a span
Instead of wrapping PropRep in a span which doesn't serve any purpose
except complying to React limits, directly return an array of the
elements we want.
This implies changes in the Rep that uses PropRep since they have to "unfold"
the PropRep elements.
We take advantage of this rewrite to simplify how we deal with commas in Rep
that have multiple children which need to be delimited by commas.
Also fixed some case where we were doing unnecessary work in TINY mode.