Modify how we indicate there is more to see in an object.
Basically, for the most case we switched from "more…" to "…" only.
We also re-use the ellipsis in TINY mode for arrays and objects.
Now empty arrays and objects in TINY mode should render respectively
as "[]" and "{}", and non-empty arrays and objects "[…]" and "{…}".

This should indicate more clearly and in a more consistent way that
there are hidden properties/items in what we render, and encourage the
user to expand the object/array to check what it really contains.

Doing this work was also a great opportunity to be more consistent through
all the reps, and mostly have the same output for RDP and "plain" objects
for the same test cases.