Display getters and setters in Reps and in the ObjectInspector.
This PR adds a `Accessor` rep which handle object with getter and/or setter.
I choose a dedicated Rep because the plan is to make it evolve in the future,
for example to be able to evaluate a property with a getter.
This also adds pseudo properties on the ObjectInspector so the getter and the
setter can also be inspected, like we could in the VariableView.