Fix array output when items are not in the preview.
We can have cases where we don't have the array items in the preview,
for example in a packet of a Promise resolved with an array, but only
the length of the array.
In this case, it's better to tell the user there's actually something
in the array (showing ellipsis).