Handle exceptions in network request and worker utilities
This changes the network request code and the worker handler code to
catch exceptions and report them to the caller via their promise
results.  It also changes the network request code to provide an error
message when the XHR's statusText is empty; this can happen when an
invalid URL is used.