Add a MapEntry Rep.
When we fetch entries from the server, we get a grip
with a "mapEntry" type that we weren't handling previously.
This patch adds a Rep to deal with those kind of grips and
harmonize the way we render map entries (even in the map preview),
to match what we had previously in the VariableView.