Adapt Launchpad for entries.
This provides the getObjectEntries and loadObjectEntries to the
ObjectInspector used in the Result component.
A new "maps & sets" quick link is added as well to be able to
quickly see maps and states of different size.
We take this as a opportunity to add redux logger to it so we
can see what actions are fired when.