Fix bugs for grips without preview property.
In some cases, we can have a grip without a preview property (e.g.
for the following object `[{}]`). Because the grip was only supporting
object with preview property, the grip was rendered with the default
rep (i.e. Object), and thus could be incorrect.
This patch fixes this and adds some tests to make sure we handle this
case as expected.