Change default properties node styling.
First, we add the missing rule for the default-property class.
This patch changes how we style default properties node in
objects. Previously, we were looking for properties with the
same name as a window ownProperty. This could be dangerous
since you can have properties with the same name as one in
Window, but not necessarely under the window object.
Also, this can be weird to dim all the window default properties
when you expanded the node. If you did so, you are probably
interested in seeing what's inside.
Taking this into account, this patch applies the default-property class
to the default properties node and __proto__ ones, but only when they are
collapsed.