Do not put a rejected promise into sourceMapRequests cache
My previous patch in this area was mistaken - it correctly allowed
errors from _fetchSourceMap to be propagated to the caller, but it
also cached the rejected promise in sourceMapRequests, resulting in
the same error being reported via other APIs.  This patch fixes the
oversight by caching a non-rejecting promise; this results in the
other APIs keeping their previous behavior.