Add a way to load and display entries.
This adds a loadObjectEntries as well as a getObjectEntries function as props,
which will be used to fetch the entries of an object.
The loadObjectEntries is called only when a custom <entries> node is expanded
by the user (if we don't already have all the entries in the preview object).