Handle Proxy target and handler in ObjectInspector.
Previously, when the user expanded a Proxy object, we loaded the properties
of the target, which is only half the information the user might want to see.
Now we build 2 ndoes, target and handler, that can be expanded as well so
the user can inspect the Proxy object properly.
Some tests were added to make sure we handle proxies as expected, and the
setExpanded function was refactored a bit to be more readable.