Don't add root actors to the state of the ObjectInspector.
The actors property on the ObjectInspector state keep track of all the actors that need to be
released when we're done with the component. The problem of releasing the root actor is that
if the comsumer of the OI was simply hidden, and then shown again, the user would not be
able to expand the objectInspector.
Here we don't track root actors so the consumer handle the lifetime of the root actors it handles.
This causes no issues for non-root actors, since the ObjectInspector is unmounted, if the user
want to expand it again after a hide/show cycle again, we'll end up creating new ObjectClient (and
actors thus).