Wrap provided objectClient function calls in try/catch blocks.
Since we don't know what the ObjectClient we get looks like, it might happens that it doesn't have
the method we try to call (e.g. connecting to an older server, you will miss the enumSymbols function).
Wrapping everything in try/catch blocks mitigate the risk of crashing the ObjectInspector (in the beforementioned
case, everything will be fine except we wouldn't get symbol properties, which seems acceptable).