Unvirtualize the Tree component.
We were having an onScroll event that was making the console lag
if you had a large number of ObjectInspector in it.
None of the consumer of this Tree component take advantage of this
so let's remove it.
We might want to reintroduce some virtualization later, but maybe using
the IntersectionObserver instead, since we can have variable height items.
Also this PR breaks keyboard navigation, as if you navigate to an
off-canvas item you won't be scrolled. I think this was already broken for
variable height items though. This should be fixed in another PR
with keeping in mind that this tree is not meant to be used as the single
component of a container. At the moment, every consumer disable the
keyboard navigation, so we don't put anyone at risk.