Guard nodeHasAllEntriesInPreview.
Fixes #681.

Some map can have a preview object which contains only a size property.
In such case, the ObjectInspector was throwing in nodeHasAllEntriesInPreview,
since it expected to have at least an entries or items property.
We now return early on the function if the preview object does not have an
entries nor an item property.