Compute sourceRoot when only relative URLs are seen
Currently _setSourceMapRoot bails if the source map contains all the
sources.  However, this is not valid according to the source map spec;
and the debugger doesn't work properly with the returned URLs.  On the
other hand, some source maps have non-URLs like "webpack:/bootstrap",
which we do want to treat as absolute.  So, this patch works around
the problem by examining the URLs in a mildly lax way.
Fixes #356