Override console.error in create-object-client.
In this test, we check that the ObjectInspector does not throw if
the create object client has not the method we expect.
If the objectClient is faulty, we do use console.error to warn the
developer though, and those console calls are spamming the terminal
during the test, which can make you think that some tests are failing.
Here we override console.error so we have a saner output.