Add a button to jump to a function definition.
Fixes #640.

The Function rep would accept an onViewSourceInDebugger prop, and the
icon would be shown when we do have this props as well as a location with
a non-null url (so we don't show it on prototypes for example).
Clicking on the button would call the prop with the location object.
A test is added to make sure we handle the different possible cases.