Fix style for non-expandable root nodes in Tree component.
Fixes #886.

If a Tree contained multiple roots, some expandables, some not, the
non-expandable node weren't aligned to the expanded one.
This is because we align non-expandable items adding a margin to the
last tree-indent element of the node.
However, non-expandable root do not have any tree-indent element child,
so they were missing a margin.
This patch adds a specific margin-inline-start for non-expandable root nodes.
The multiple root tree story is modified to also include non-expandable root
nodes.