Add DocumentType rep.
Fixes #945.

Previously it was rendered by the Grip Rep and was causing issue
because the packet does not have a ownProperties prop.
Adding a dedicated Rep, with similar output as in the markup view fixes
the issue. Stubs and tests are added as well.
And to be safe, we default ownProperties to an empty object in the Grip
rep so we don't have errors anymore.