Don't inline SVG images in devtools-reps.
This causes warning in the browser console AND can cause performance issue.
Using those images as background images means we can remove InlineSVG dependency
and helpers.
It also means we need to have a postCSS config to be able to transform those
images URL for development in launchpad, and also for the copy-assets command.