Re-create loadItemProperties function which is used by the debugger.
The function was deleted in https://github.com/devtools-html/devtools-core/commit/7c538787e1f19c07e868192268097f6c8845feb2\#diff-1d5a77fb4120fb410c4af9fa9a4eb936
because it wasn't used in devtools-reps code.
But it was exported and the debugger does use it for
the Preview popup.
So we re-create it and make sure the similar action use
it as well so the ObjectInspector and debugger code path
don't diverge.