Don't use ObjectInspector for Error grips.
Fixes #960.
Since the ErrorRep already provides all the error informations
in a nice way (message and stacktrace), there's no benefits
being able to inspect the object. Furthermore, when in the console,
throwing an Error object gives you a double arrow (first to expand
the stack we get from the exception, second is the objectInspector
arrow), which is clearly not ideal.
Using Rep directly solves these issues.