Fix Error Rep with longString stacktrace.
This was throwing because we are calling split on preview.stack,
as we thought having a string. But if this is a longString, we get
an object and thus we have to pick longString's initial value.
This can be considered as a hotfix since we are not properly rendering
the whole stacktrace in this case. This will be handled in a separate
issue, maybe directly on the server side.