Use symbols for nodes path.
Fixes #1001, #1002 & #1003.

The path property on a node is very important as it serve
as an identifier to know if a node is expanded, and what
its properties are. Currently, this path was a simple string
which could led to 2 nodes having the same path, which could
have unwanted effect on user interaction.
To avoid this, the path is now a symbol, which has the advantage
of being unique.
Because of this change, we needed to adapt our code so we always
have the same path reference in the state.
With this patch, we'll always dispatch the nodePropertiesLoaded action,
even if the node don't have any properties. We'll store null in the state,
so we can then check if a node was properly handled.
As a result, we change the shouldComponentUpdate of the component
to reflect this new behavior.
In order to have always the same reference, we always add created node
to the cacheNodes property, which we can do because of the change in
loadedProperties.
This should also give us some performance improvements as we don't have
to recreate some nodes twice.
We take this as an opportunity to change the signature of the createNode
function to be more explicit (it now expect an object of options), and
we also create the path of the node from its parent when the path option
is not supplied.