Fix expanding node with selected text.
Fixes #1017.

Since clicking on the arrow does NOT unselect potential selected
text, we need to check the target of the click and see if it is
the node arrow. In such case, we don't block the expansion.
We take this as an opportunity to remove now unnecessary code which
was set to deal with arrows when they were svg elements (now we use
images and don't have such issues).