Cleanup redux state when roots prop changes.
The roots prop isn't handled by the redux state, but is passed from
upper-level components (e.g. console, inspector, ...). There might be
cases when the roots prop changes (e.g., in the console, the object sidebar
is already opened, and the user choose to put another object in it).
In such cases, there are a few things that should be taken care of:
- we need to clear the node cache to start anew
- we need to clean the redux state of the old roots-related properties,
  like expandedPaths, loadedProperties, …
- we need to release the actors related to the old roots.
- we need to make sure a possible new focusedItem is handled as expected.

In order to do that, we create a rootsChanged action that's handled in
a middleware where we reinitialize the state and release actors.
We don't want to render before the state is re-initialized, so we add a
new state property, forceUpdate, that we can then use as a flag that we
set when the roots prop changes and reset when the component was updated.

This isn't perfect but I wasn't able to find a better way of handling all
those requirements.