Prevent unnecessary server call when expanding node.
When expanding a node, in some cases, we query the server to know how
much indexed, non-indexed and symbol properties we have in the object.
The issue is that we then always try to get all the properties of the
object, even if the server told us there was none.
This patch prevent this and should save a couple of client-server round-trip.