Speed up containsURL function.
The function was splitting up a string and testing if each parts was an actual URL,
until one was encountered.
The split function was showing up in some profile I did today (nothing too big, but
bigger that it should be).
Here switching to a simple regex test works and is faster to run.